RakshaNet/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
Django==4.2.8
django-crispy-forms>=2.1
crispy-bootstrap4>=1.0.2
djangorestframework>=3.14.0
Pillow>=9.0.0
python-dateutil==2.8.2
pytz==2024.1
sqlparse==0.4.4
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
